fix(db): restore shard status on failed vector index config update
