+++ adapters/repos/db/shard_init_vector_toctou_test.go
@@ -13,11 +13,19 @@ package db
 
 import (
 	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+	"runtime"
+	"strings"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	schemaConfig "github.com/weaviate/weaviate/entities/schema/config"
+	"github.com/weaviate/weaviate/entities/storagestate"
 	enthnsw "github.com/weaviate/weaviate/entities/vectorindex/hnsw"
 )
 
@@ -47,6 +55,88 @@ func TestInitTargetVector_Idempotent_DoesNotOrphanQueue(t *testing.T) {
 	assert.Same(t, q1, q2,
 		"re-initialising an existing target vector must not silently replace and "+
 			"orphan its queue; the first queue is leaked (never Dropped)")
+}
+
+// TestInitTargetVector_ShutsDownIndexWhenQueueCreationFails pins that a vector
+// index whose queue fails to be created is shut down rather than orphaned,
+// leaking its cache goroutine. The cancelled-ctx case guards that the shutdown
+// uses s.shutCtx and not the caller's (possibly cancelled) ctx.
+func TestInitTargetVector_ShutsDownIndexWhenQueueCreationFails(t *testing.T) {
+	tests := []struct {
+		name      string
+		callerCtx func() context.Context
+	}{
+		{name: "valid caller ctx", callerCtx: context.Background},
+		{
+			name: "cancelled caller ctx",
+			callerCtx: func() context.Context {
+				ctx, cancel := context.WithCancel(context.Background())
+				cancel()
+				return ctx
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// async indexing makes NewVectorIndexQueue call q.Init()
+			shardLike, _ := testShard(t, context.Background(), "VecQueueOrphan",
+				func(idx *Index) { idx.AsyncIndexingEnabled = true })
+			s := underlyingShard(t, shardLike)
+
+			const target = "orphanTarget"
+			// a file where the queue dir must go makes q.Init() fail
+			queueDir := filepath.Join(s.path(), fmt.Sprintf("%s.queue.d", s.vectorIndexID(target)))
+			require.NoError(t, os.WriteFile(queueDir, []byte("x"), 0o600))
+
+			// the HNSW cache goroutine only exits on Shutdown/Drop
+			watchers := func() int {
+				buf := make([]byte, 1<<20)
+				for {
+					n := runtime.Stack(buf, true)
+					if n < len(buf) {
+						return strings.Count(string(buf[:n]), "watchForDeletion")
+					}
+					buf = make([]byte, 2*len(buf))
+				}
+			}
+			baseline := watchers()
+
+			err := s.initTargetVector(tt.callerCtx(), target, enthnsw.NewDefaultUserConfig(), false)
+			require.Error(t, err)
+			require.ErrorContains(t, err, "cannot create index queue")
+
+			s.vectorIndexMu.RLock()
+			_, stored := s.vectorIndexes[target]
+			s.vectorIndexMu.RUnlock()
+			require.False(t, stored, "orphaned index must not be stored")
+
+			require.Eventually(t, func() bool { return watchers() <= baseline }, 5*time.Second, 20*time.Millisecond,
+				"orphaned HNSW index was not shut down (cache goroutine leaked)")
+		})
+	}
+}
+
+// TestUpdateVectorIndexConfigs_FailedInitRestoresStatus pins that a failed
+// vector-index creation does not leave the shard read-only: the goroutine that
+// restores StatusReady must run even when initTargetVector returns an error.
+func TestUpdateVectorIndexConfigs_FailedInitRestoresStatus(t *testing.T) {
+	// async indexing makes NewVectorIndexQueue call q.Init()
+	shardLike, _ := testShard(t, context.Background(), "VecQueueReadonly",
+		func(idx *Index) { idx.AsyncIndexingEnabled = true })
+	s := underlyingShard(t, shardLike)
+
+	const target = "newTarget"
+	// a file where the queue dir must go makes q.Init() fail
+	queueDir := filepath.Join(s.path(), fmt.Sprintf("%s.queue.d", s.vectorIndexID(target)))
+	require.NoError(t, os.WriteFile(queueDir, []byte("x"), 0o600))
+
+	err := s.UpdateVectorIndexConfigs(context.Background(),
+		map[string]schemaConfig.VectorIndexConfig{target: enthnsw.NewDefaultUserConfig()})
+	require.Error(t, err)
+
+	require.Eventually(t, func() bool { return s.GetStatus() == storagestate.StatusReady }, 5*time.Second, 20*time.Millisecond,
+		"shard left read-only after a failed vector index creation")
 }
 
 // underlyingShard returns the concrete *Shard behind a ShardLike, loading it if
